pulsos por galon

+++ include/variables.h
@@ -6,5 +6,5 @@ String ID_Modulo = "";
 int contador_impulsos; //Conteo de impulsos.
 int Valor_inicial = 0; //valor inicial.
 int pulsos = 0;
-int cantidad = 0;
+float cantidad = 0;
 String status = "";
